Move records endpoint into Domains Controller
